Refactor components to Typescript
