+++ src/App.css
@@ -1,12 +1,62 @@
+body {
+	--primary-color: #3bb3bd;
+	--primary-color-highlight: #4bbdc7;
+	--online-color: green;
+	--offline-color: red;
+}
+
+body {
+	margin: 3rem;
+}
+
 .online {
-	color: green;
+	color: var(--online-color);
+	margin-left: 1rem;
 }
 
 .offline {
-	color: red;
+	color: var(--offline-color);
+	margin-left: 1rem;
 }
 
 .controlBtn {
-	width: 100px;
-	height: 40px;
+	border-radius: 1rem;
+	border: none;
+	cursor: pointer;
+	font-size: 1.5rem;
+	line-height: 1;
+	margin: 0.5rem;
+	max-width: 100%;
+	min-height: 44px;
+	outline: 0;
+	overflow: hidden;
+	padding: 1rem 1.5rem;
+	text-align: center;
+	text-decoration: none;
+	text-overflow: ellipsis;
+	transition: background-color .1s;
+	white-space: normal;
+	position: relative;
 }
+
+.controlBtn:active {
+	top: 2px;
+}
+
+.controlBtn:first-child {
+	background-color: var(--primary-color);
+}
+
+.controlBtn:first-child:hover {
+	background-color: var(--primary-color-highlight);;
+}
+
+.controlBtn:nth-child(2) {
+	background-color: #edeff0;
+	color: #384047;
+}
+
+.controlBtn:nth-child(2):hover {
+	background-color: rgba(237,239,240,.7);
+}
+
